+++ R/plotBiomassObservedVsModel.R
@@ -193,7 +193,7 @@ plotBiomassObservedVsModel.MizerParams <- function(object, species = NULL, ratio
 plotBiomassObservedVsModel.MizerSim <- function(object, species = NULL, ratio = FALSE,
                                       log_scale = TRUE, return_data = FALSE, 
                                       labels = TRUE, show_unobserved = FALSE, ...) {
-    params <- setInitialValues(object@params, object)
+    params <- finalParams(object)
     plotBiomassObservedVsModel(params, species = species, ratio = ratio,
                                log_scale = log_scale, return_data = return_data,
                                labels = labels, show_unobserved = show_unobserved)